data_scrapper/requirements.txt
requests
mysql-connector-python
schedule
pytz
pandas
python-dotenv
